pan_tilt_serial_project/backend/requirements.txt
pyserial>=3.5
flask>=2.0
flask-socketio>=5.0
python-socketio>=5.0
